rename NewZkSyncEraDiff -> ZkSyncEraDiff
